requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.6.1; python_full_version >= '3.6.2'
click==8.1.3; python_version >= '3.7'
colorama==0.4.5; platform_system == 'Windows'
fastapi==0.85.0
h11==0.13.0; python_version >= '3.6'
idna==3.4; python_version >= '3.5'
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3; python_version >= '3.6'
pillow==9.2.0
pydantic==1.10.2; python_version >= '3.7'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pytesseract==0.3.10
sniffio==1.3.0; python_version >= '3.7'
starlette==0.20.4; python_version >= '3.7'
typing-extensions==4.3.0; python_version < '3.10'
uvicorn==0.18.3
